Добавлен поиск по названию фильма, внесены правки в структуру и модель

+++ src/main/java/ru/yandex/practicum/filmorate/service/GenreService.java
@@ -19,7 +19,7 @@ public class GenreService {
         return genreStorage.getAllGenres();
     }
 
-    public Genre getGenreById(Long id) {
+    public Genre getGenreById(Integer id) {
         Genre genre = genreStorage.getGenreById(id);
         if (genre == null) {
             throw new GenreNotFoundException("Жанр с id " + id + " не найден");
